Dev into main

+++ JourneyService/src/main/java/com/skipass/journeyservice/business/impl/JourneyServiceImpl.java
@@ -1,7 +1,6 @@
 package com.skipass.journeyservice.business.impl;
 
 import com.skipass.journeyservice.business.JourneyService;
-import com.skipass.journeyservice.domain.Journey;
 import com.skipass.journeyservice.domain.JourneyType;
 import com.skipass.journeyservice.persistance.JourneyEntity;
 import com.skipass.journeyservice.persistance.JourneyRepository;
@@ -18,9 +17,8 @@ import java.util.Random;
 public class JourneyServiceImpl implements JourneyService {
 
     private final JourneyRepository journeyRepository;
-
     @Override
-    public Journey createJourney(String authorUsername) {
+    public JourneyEntity createJourney(String authorUsername) {
         Random random = new Random();
         JourneyType[] types = JourneyType.values();
         JourneyType type = types[random.nextInt(types.length)];
@@ -29,7 +27,7 @@ public class JourneyServiceImpl implements JourneyService {
         int totalPasses;
         switch (type) {
             case FULL_DAY:
-                totalKilometers =50 + random.nextInt(150);
+                totalKilometers = 50 + random.nextInt(150);
                 totalPasses = 8 + random.nextInt(30);
                 break;
             case MORNING:
@@ -54,10 +52,11 @@ public class JourneyServiceImpl implements JourneyService {
         int slowestRun = fastestRun + random.nextInt(10);
 
         LocalDate date = LocalDate.now();
-
-        Journey journey = Journey.builder().authorUsername(authorUsername).date(date).fastest(fastestRun)
+        JourneyEntity journey = JourneyEntity.builder().authorUsername(authorUsername).date(date).fastest(fastestRun)
                 .type(type).slowest(slowestRun).totalKm(totalKilometers).totalPasses(totalPasses).build();
-        return journey;
+       var returned = journeyRepository.save(journey);
+
+        return returned;
     }
 
     @Override
@@ -77,7 +76,10 @@ public class JourneyServiceImpl implements JourneyService {
 
     @Override
     public void deleteJourneyById(long journeyId) {
-        journeyRepository.deleteById(journeyId);
+        Optional<JourneyEntity> journey = journeyRepository.findById(journeyId);
+        if(journey.isPresent()) {
+            journeyRepository.deleteById(journeyId);
+        }
     }
 
     @Override
